Refactor: change the validation of target for the condition
Some conditions don't have target.
For example, getting a badge for receiving all other achievements, or for absolutely all passed tests.
In such cases, the badge must be saved with a null value in the "target" field.
Otherwise, the last value selected on the form will be saved in this field.
This may negatively affect the retention of the "condition+target" index.

Splitting an array CONDITIONS into two is a temporary solution.
In the future, method need_target? it will be implemented better.